Adding days 6, 7, and 8

+++ 2024/AdventOfCode/Challenges/Challenge.cs
@@ -1,6 +1,6 @@
 ﻿namespace AdventOfCode2024;
 public abstract class Challenge
 {
-    public abstract int Solution1(string inputPath);
-    public abstract int Solution2(string inputPath);
+    public abstract long Solution1(string inputPath);
+    public abstract long Solution2(string inputPath);
 }
